+++ geomstats/numerics/optimization.py
